Clean up animation variables
