+++ client/src/app/Button/Button.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components/macro'
 import PropTypes from 'prop-types'
 import { motion } from 'framer-motion'
+import { clickAnimation } from '../../services/animationVariants'
 
 Button.propTypes = {
   onClick: PropTypes.func.isRequired,
@@ -17,11 +18,6 @@ export default function Button({
   isNoCircle,
   isSmall,
 }) {
-  const clickAnimation = {
-    rest: { scale: 1 },
-    pressed: { scale: 0.9 },
-  }
-
   return (
     <motion.div variants={clickAnimation} initial="rest" whileTap="pressed">
       <ButtonStyled
